requirements.txt
playwright
openai
PyYAML
chromadb
aiofiles
httpx